slub: No SLUB merging for disable_higher_order_debug
The SLUB debug option "o" is a feature to prevent increasing the size of
the slab size order when metadata is used. This is not compatible with
the find_mergeable() search loop because the loop does not take into
account the change in flags and change in slab size when
disable_higher_order_debug is triggered.

Therefore, inhibit SLUB merging when disable_higher_order_debug
is enabled.

This has been implemented by creating a SLUB cache specific function
called slub_unmergeable() to test SLUB specific unmergeable conditions for
SLUB caches.

Signed-off-by: Dean Jenkins <Dean_Jenkins@mentor.com>