slub: Modify find_mergeable() to use prepare_size()
Use prepare_size() in find_mergeable() to calculate the predicted
slab size of the requested SLUB cache. This means the size of the
requested SLUB cache and the size used to create a SLUB cache both
use the same size calculation functions.

This simplfies the mergeable search loop as only the calculated slab
size of the requested SLUB cache needs to be equal to the s->size
slab size of an existing SLUB cache.

Therefore, remove the now redundant checks from the search loop.

Signed-off-by: Dean Jenkins <Dean_Jenkins@mentor.com>