slab: Modify find_mergeable() to use prepare_size()
Use prepare_size() in find_mergeable() to calculate the predicted
slab size of the requested SLAB or SLUB cache. This means the slab
size of the requested cache and the slab size used to create a slab
cache both use the same size calculation functions.

This simplfies the mergeable search loop as only the calculated slab
size of the requested slab cache needs to be equal to the s->size
slab size of an existing cache.

Therefore, remove the now redundant checks from the search loop.

Signed-off-by: Dean Jenkins <Dean_Jenkins@mentor.com>