slub: Maximise object_size
To make SLUB merging simpler, maximise the object_size when the SLUB
cache is created so that requested SLUB caches of the same slab size
but with a smaller or equal object_size can be merged together.

Use an allow_maximise function parameter to control whether the
object_size is allowed to be maximised. This is needed to prevent
the object_size of an existing SLUB cache to be modified such as
during modification of run-time SLUB debug flags. Otherwise, a
reduction in metadata caused by changes in the flags can lead
to the object_size being unnecessarily increased. Also, subsequent
run-time changes to SLUB debug flags could cause the slab size
to be undesirably increased because the object_size was previously
increased.

Remove the now redundant object_size increasing code
from __kmem_cache_alias().

For consistency, use slab_unmergeable() to determine whether
a SLUB cache is mergeable.

Signed-off-by: Dean Jenkins <Dean_Jenkins@mentor.com>